IEI-175871 DCM4CHE3 does not support elements> 2GB in length
- Cherry picked commits 28a3a5b2a02eb96d920d80dab7bc670924a0f7b9 and 930f9ac849f2a12a5db33c8adf2355ac36c30203 from master branch
- Full manual merge for BulkData.java and partial manual merge for DicomInputStream.java, the rest were automatic.
- Additional fixes in CloseableImageInputStreamAdapter and BulkDataDicomImageAccessor since these don't exist in master

+++ dcm4che-core/src/main/java/org/dcm4che3/io/stream/BulkURIImageInputStream.java
@@ -52,11 +52,11 @@ public class BulkURIImageInputStream extends ImageInputStreamImpl {
 
     private final ImageInputStream iis;
     private final long offset;
-    private final int length;
+    private final long length;
 
 
 
-    public BulkURIImageInputStream(ImageInputStream iis, long offset, int length) throws IOException {
+    public BulkURIImageInputStream(ImageInputStream iis, long offset, long length) throws IOException {
         this.iis = iis;
         this.offset = offset;
         this.length = length;
